+++ controller/user/edit-user.php
@@ -3,13 +3,13 @@
     ini_set('display_startup_errors', 1);
     error_reporting(E_ALL);
 
-    include 'validation/validation.php';
-    include 'Database/Database.php';
+    include '../../validation/validation.php';
+    include '../../models/user.php';
 
     session_start();
    
     if(empty($_SESSION) || $_SESSION['role'] !== 1){
-        header("Location:login-form.php");
+        header("Location:../../views/auth/login-form.php");
     }
     // call validation and extract needed data
     $validations = validate();
@@ -18,9 +18,10 @@
     $profile_name = $validations["profile_name"];
     $profile_tmp = $validations["profile_tmp"];
 
-    // redirect to add form with errors and old data
+    // redirect to edit form with errors and old data
     if($formerrors !== "[]"){
-        $redirect_url = "Location:add-userForm.php?errors={$formerrors}";
+        $id = $_GET['id'];
+        $redirect_url = "Location:../../views/user/edit-form.php?errors={$formerrors}&id={$id}";
         if ($oldvalues !== "[]"){
             $oldvalues = json_encode($oldvalues);
             $redirect_url .="&old={$oldvalues}";
@@ -31,29 +32,26 @@
         // uploading image
         if($profile_name != ""){
             $imagespath = $_GET["imgPath"];
-            unlink($imagespath);
+            if(file_exists( $imagePath))  unlink($imagespath);
             sys_get_temp_dir();
-            move_uploaded_file($profile_tmp,"images/users/{$profile_name}");
-            $imagespath = "images/users/{$profile_name}";
+            move_uploaded_file($profile_tmp,"../../assets/images/users/{$profile_name}");
+            $imagespath = "../../assets/images/users/{$profile_name}";
         }
 
         // update data 
         try {
             // connect to database 
             $database = new Database();
-            $db = $database -> connect();
-
             $id = $_GET['id'];
             
-           
             if($imagespath)        
                 $database -> update("User", $id,"username", "email","password", "room", "`profile-pic`",
-                $_POST['name'], $_POST['email'], $_POST['password'],$_POST['room'], $imagespath);
+                $_POST['username'], $_POST['email'], $_POST['password'],$_POST['room'], $imagespath);
             else 
-                $database -> update( "users", $id,"name", "email","password", "room", 
-                $_POST['name'], $_POST['email'], $_POST['password'],$_POST['room']);
+                $database -> update( "users", $id,"username", "email","password", "room", 
+                $_POST['username'], $_POST['email'], $_POST['password'],$_POST['room']);
            
-            header("Location:users-table.php");
+            header("Location:../../views/user/users-table.php");
         } catch (Exception $e) {
             var_dump( $stmt->errorInfo());
             var_dump($e);
